test + init instruction

+++ programs/app/src/state/Counter.rs
@@ -1,7 +1,8 @@
 use anchor_lang::prelude::*;
 
+#[account] //self-reminder- this is a marking for an 'ANCHOR' account -> macro for impl serializer
 pub struct Counter {
     pub authority: Pubkey,
-    pub count: u64,
+    pub count: u8,
     pub bump: u8,
 }
